--- a/idx/memory/memory.go
+++ b/idx/memory/memory.go
@@ -344,21 +344,21 @@ func (m *MemoryIdx) find(orgId int, pattern string) ([]*Node, error) {
 	// for a query like foo.bar.baz, pos is 2
 	// for a query like foo.bar.* or foo.bar, pos is 1
 	// for a query like foo.b*.baz, pos is 0
-	pos := len(nodes) - 1
+	pos := len(nodes)
 	for i := 0; i < len(nodes); i++ {
 		if strings.ContainsAny(nodes[i], "*{}[]?") {
 			log.Debug("memory-idx: found first pattern sequence at node %s pos %d", nodes[i], i)
-			pos = i - 1
+			pos = i
 			break
 		}
 	}
 	var startNode *Node
-	if pos == -1 {
+	if pos == 0 {
 		//we need to start at the root.
 		log.Debug("memory-idx: starting search at the root node")
 		startNode = tree.Items[""]
 	} else {
-		branch := strings.Join(nodes[0:pos+1], ".")
+		branch := strings.Join(nodes[0:pos], ".")
 		log.Debug("memory-idx: starting search at branch %s", branch)
 		startNode, ok = tree.Items[branch]
 		if !ok {
@@ -367,23 +367,21 @@ func (m *MemoryIdx) find(orgId int, pattern string) ([]*Node, error) {
 		}
 	}
 
-	if pos == len(nodes)-1 {
-		// startNode is the leaf we want.
-		log.Debug("memory-idx: pattern %s was a specific branch/leaf name.", pattern)
-		results = append(results, startNode)
-		return results, nil
-	}
-
 	children := []*Node{startNode}
-	for pos < len(nodes) {
-		pos++
-		if pos == len(nodes) {
-			log.Debug("memory-idx: reached pattern length at node pos %d. %d nodes matched", pos, len(children))
-			for _, c := range children {
-				results = append(results, c)
+	for i := pos; i < len(nodes); i++ {
+		p := nodes[i]
+
+		matchAll := "*" == p
+		var regex *regexp.Regexp
+		if !matchAll && strings.ContainsAny(nodes[i], "*{}[]?") {
+			r, err := regexp.Compile(toRegexp(p))
+			if err != nil {
+				log.Debug("memory-idx: regexp failed to compile. %s - %s", p, err)
+				return nil, err
 			}
-			continue
+			regex = r
 		}
+
 		grandChildren := make([]*Node, 0)
 		for _, c := range children {
 			if !c.HasChildren() {
@@ -391,10 +389,28 @@ func (m *MemoryIdx) find(orgId int, pattern string) ([]*Node, error) {
 				// expecting a branch
 				continue
 			}
-			log.Debug("memory-idx: searching %d children of %s that match %s", len(c.Children), c.Path, nodes[pos])
-			matches, err := match(nodes[pos], c.Children)
-			if err != nil {
-				return results, err
+			log.Debug("memory-idx: searching %d children of %s that match %s", len(c.Children), c.Path, nodes[i])
+			matches := make([]string, 0)
+			if regex != nil {
+				for _, c := range c.Children {
+					if regex.MatchString(c) {
+						log.Debug("memory-idx: %s regex matches %s", c, regex.String())
+						matches = append(matches, c)
+					}
+				}
+			} else if (matchAll) {
+				log.Debug("Matching all children")
+				for _, c := range c.Children {
+					matches = append(matches, c)
+				}
+			} else {
+				for _, c := range c.Children {
+					if c == p {
+						log.Debug("memory-idx: %s matches %s", c, p)
+						matches = append(matches, c)
+						break
+					}
+				}
 			}
 			for _, m := range matches {
 				newBranch := c.Path + "." + m
@@ -411,43 +427,11 @@ func (m *MemoryIdx) find(orgId int, pattern string) ([]*Node, error) {
 		}
 	}
 
-	return results, nil
-}
-
-func match(pattern string, candidates []string) ([]string, error) {
-	var patterns []string
-	if strings.ContainsAny(pattern, "{}") {
-		patterns = expandQueries(pattern)
-	} else {
-		patterns = []string{pattern}
+	log.Debug("memory-idx: reached pattern length. %d nodes matched", pos, len(children))
+	for _, c := range children {
+		results = append(results, c)
 	}
 
-	results := make([]string, 0)
-	for _, p := range patterns {
-		if strings.ContainsAny(p, "*[]?") {
-			p = strings.Replace(p, "*", ".*", -1)
-			p = strings.Replace(p, "?", ".?", -1)
-			p = "^" + p + "$"
-			r, err := regexp.Compile(p)
-			if err != nil {
-				log.Debug("memory-idx: regexp failed to compile. %s - %s", p, err)
-				return nil, err
-			}
-			for _, c := range candidates {
-				if r.MatchString(c) {
-					log.Debug("memory-idx: %s matches %s", c, p)
-					results = append(results, c)
-				}
-			}
-		} else {
-			for _, c := range candidates {
-				if c == p {
-					log.Debug("memory-idx: %s is exact match", c)
-					results = append(results, c)
-				}
-			}
-		}
-	}
 	return results, nil
 }
 
@@ -457,7 +441,7 @@ func (m *MemoryIdx) List(orgId int) []idx.Archive {
 	defer m.RUnlock()
 	orgs := []int{-1, orgId}
 	if orgId == -1 {
-		log.Info("memory-idx: returing all metricDefs for all orgs")
+		log.Info("memory-idx: returning all metricDefs for all orgs")
 		orgs = make([]int, len(m.Tree))
 		i := 0
 		for org := range m.Tree {
@@ -639,42 +623,35 @@ func (m *MemoryIdx) Prune(orgId int, oldest time.Time) ([]idx.Archive, error) {
 	return pruned, nil
 }
 
-// filepath.Match doesn't support {} because that's not posix, it's a bashism
-// the easiest way of implementing this extra feature is just expanding single queries
-// that contain these queries into multiple queries, who will be checked separately
-// and whose results will be ORed.
-func expandQueries(query string) []string {
-	queries := []string{query}
-
+func toRegexp(pattern string) string {
 	// as long as we find a { followed by a }, split it up into subqueries, and process
 	// all queries again
 	// we only stop once there are no more queries that still have {..} in them
-	keepLooking := true
-	for keepLooking {
-		expanded := make([]string, 0)
-		keepLooking = false
-		for _, query := range queries {
-			lbrace := strings.Index(query, "{")
-			rbrace := -1
-			if lbrace > -1 {
-				rbrace = strings.Index(query[lbrace:], "}")
-				if rbrace > -1 {
-					rbrace += lbrace
-				}
-			}
+	p := pattern
 
-			if lbrace > -1 && rbrace > -1 {
-				keepLooking = true
-				expansion := query[lbrace+1 : rbrace]
-				options := strings.Split(expansion, ",")
-				for _, option := range options {
-					expanded = append(expanded, query[:lbrace]+option+query[rbrace+1:])
-				}
-			} else {
-				expanded = append(expanded, query)
+	pos := 0
+	for pos < len(p) {
+
+		lbrace := strings.Index(p[pos:], "{")
+		rbrace := -1
+		if lbrace > -1 {
+			lbrace += pos
+			rbrace = strings.Index(p[lbrace:], "}")
+			if rbrace > -1 {
+				rbrace += lbrace
 			}
 		}
-		queries = expanded
+
+		if lbrace == -1 || rbrace == -1 {
+			break
+		}
+		p = p[:lbrace] + "(" + strings.Replace(p[lbrace+1:rbrace], ",", "|", -1) + ")" + p[rbrace+1:]
+		pos = rbrace + 1
+
 	}
-	return queries
+
+	p = strings.Replace(p, "*", ".*", -1)
+	p = strings.Replace(p, "?", ".?", -1)
+	p = "^" + p + "$"
+	return p
 }