--- a/idx/memory/memory.go
+++ b/idx/memory/memory.go
@@ -606,19 +606,33 @@ func getMatcher(path string) (func([]string) []string, error) {
 		}, nil
 	}
 
+	var patterns []string
+	if strings.ContainsAny(path, "{}") {
+		patterns = expandQueries(path)
+	} else {
+		patterns = []string{path}
+	}
+
 	// Convert to regex and match
-	if strings.ContainsAny(path, "*{}[]?") {
-		r, err := regexp.Compile(toRegexp(path))
-		if err != nil {
-			log.Debug("memory-idx: regexp failed to compile. %s - %s", path, err)
-			return nil, err
+	if strings.ContainsAny(path, "*[]?") {
+		regexes := make([]*regexp.Regexp, 0, len(patterns))
+		for _, p := range patterns {
+			r, err := regexp.Compile(toRegexp(p))
+			if err != nil {
+				log.Debug("memory-idx: regexp failed to compile. %s - %s", p, err)
+				return nil, err
+			}
+			regexes = append(regexes, r)
 		}
+
 		return func(children []string) []string {
-			matches := make([]string, 0)
-			for _, c := range children {
-				if r.MatchString(c) {
-					log.Debug("memory-idx: %s regex matches %s", c, r.String())
-					matches = append(matches, c)
+			var matches []string
+			for _, r := range regexes {
+				for _, c := range children {
+					if r.MatchString(c) {
+						log.Debug("memory-idx: %s =~ %s", c, r.String())
+						matches = append(matches, c)
+					}
 				}
 			}
 			return matches
@@ -627,43 +641,62 @@ func getMatcher(path string) (func([]string) []string, error) {
 
 	// Match a particular value
 	return func(children []string) []string {
-		for _, c := range children {
-			if c == path {
-				log.Debug("memory-idx: %s matches %s", c, path)
-				return []string{c}
+		var results []string
+		for _, p := range patterns {
+			for _, c := range children {
+				if c == p {
+					log.Debug("memory-idx: %s matches %s", c, p)
+					results = append(results, c)
+					break
+				}
 			}
 		}
-		return []string{}
+		return results
 	}, nil
 }
 
-func toRegexp(pattern string) string {
+// We don't use filepath.Match as it doesn't support {} because that's not posix, it's a bashism
+// the easiest way of implementing this extra feature is just expanding single queries
+// that contain these queries into multiple queries, which will be checked separately
+// and the results of which will be ORed.
+func expandQueries(query string) []string {
+	queries := []string{query}
+
 	// as long as we find a { followed by a }, split it up into subqueries, and process
 	// all queries again
 	// we only stop once there are no more queries that still have {..} in them
-	p := pattern
+	keepLooking := true
+	for keepLooking {
+		expanded := make([]string, 0)
+		keepLooking = false
+		for _, query := range queries {
+			lbrace := strings.Index(query, "{")
+			rbrace := -1
+			if lbrace > -1 {
+				rbrace = strings.Index(query[lbrace:], "}")
+				if rbrace > -1 {
+					rbrace += lbrace
+				}
+			}
 
-	pos := 0
-	for pos < len(p) {
-
-		lbrace := strings.Index(p[pos:], "{")
-		rbrace := -1
-		if lbrace > -1 {
-			lbrace += pos
-			rbrace = strings.Index(p[lbrace:], "}")
-			if rbrace > -1 {
-				rbrace += lbrace
+			if lbrace > -1 && rbrace > -1 {
+				keepLooking = true
+				expansion := query[lbrace+1 : rbrace]
+				options := strings.Split(expansion, ",")
+				for _, option := range options {
+					expanded = append(expanded, query[:lbrace]+option+query[rbrace+1:])
+				}
+			} else {
+				expanded = append(expanded, query)
 			}
 		}
-
-		if lbrace == -1 || rbrace == -1 {
-			break
-		}
-		p = p[:lbrace] + "(" + strings.Replace(p[lbrace+1:rbrace], ",", "|", -1) + ")" + p[rbrace+1:]
-		pos = rbrace + 1
-
+		queries = expanded
 	}
+	return queries
+}
 
+func toRegexp(pattern string) string {
+	p := pattern
 	p = strings.Replace(p, "*", ".*", -1)
 	p = strings.Replace(p, "?", ".?", -1)
 	p = "^" + p + "$"