I am testing out metrictank for usage with some timeseries with high cardinality nodes. I noticed that this caused some find operations to really lag.

I decided to look through the code to see if there was a way to speed up the find behavior. I noticed a couple of things that looked like they could be adjusted:
1. Different code path for {a,b,c} lists versus other special chars. I turned these into (a|b|c) to simplify the code flow. This did not really change the performance of the benchmarks
2. Break loop early when doing exact string match. No need to continue looking for matches and adding nodes at that point.
3. Compile the regex once for all queued children. This means that even if we have 100k children queued up in the BFS, we only compile the regex once, instead of once per child node.
4. Special case '*' and avoid regex altogether. This ended up being a pretty huge improvement.


*WARNING* : I have never written go before, so this PR might need some extra attention. For instance, I initially created an interface to simplify the code (Matcher, with RegexMatcher, ExactMatcher and AllMatcher). Turns out whatever I did just made it WAAAY slower. I still liked the code that way, but I think it needs a go experts eye to help adjust that.

Benchmarks on a 2-core 12GB RAM ubuntu-16.04 VM (dedicated):
```
benchmark                      old ns/op     new ns/op     delta
BenchmarkFind-2                257933        148660        -42.36%
BenchmarkConcurrent4Find-2     144912        87049         -39.93%
BenchmarkConcurrent8Find-2     128865        83874         -34.91%

benchmark                      old allocs     new allocs     delta
BenchmarkFind-2                1110           658            -40.72%
BenchmarkConcurrent4Find-2     1111           658            -40.77%
BenchmarkConcurrent8Find-2     1111           658            -40.77%

benchmark                      old bytes     new bytes     delta
BenchmarkFind-2                55552         30501         -45.09%
BenchmarkConcurrent4Find-2     55601         30500         -45.14%
BenchmarkConcurrent8Find-2     55618         30493         -45.17%
```

And on my 8-core mac (running bunches of other things)
```
benchmark                      old ns/op     new ns/op     delta
BenchmarkFind-8                181536        108696        -40.12%
BenchmarkConcurrent4Find-8     73666         45098         -38.78%
BenchmarkConcurrent8Find-8     66373         51848         -21.88%

benchmark                      old allocs     new allocs     delta
BenchmarkFind-8                1111           658            -40.77%
BenchmarkConcurrent4Find-8     1111           658            -40.77%
BenchmarkConcurrent8Find-8     1111           658            -40.77%

benchmark                      old bytes     new bytes     delta
BenchmarkFind-8                55620         30502         -45.16%
BenchmarkConcurrent4Find-8     55621         30497         -45.17%
BenchmarkConcurrent8Find-8     55628         30495         -45.18%
```