@shanson7 2017-07-13
Rebased against current master, changed logic to use closures to encapsulate matching functionality. This cut out a few more allocs, and removes the clutter of matching logic from the BFS logic.