@shanson7 2017-07-17
I commented out the `log.Debug` statements (as @Dieterbe noted they are polluting the results and will hopefully eventually not be evaluated in production when the logging library is replaced).

I tested 3 basic schemes:
1. This PR as-is 
2. Expanding just non-regex queries that contain {} and exact matching them
3. Expanding all {}, then either compiling them if needed, or exact matching

I tested using only these 4 patterns (so that non expanding patterns didn't wash out the results):
```
{Pattern: "collectd.{dc1,dc50}.host960.disk.disk1.disk_ops.*", ExpectedResults: 2},
{Pattern: "*.dc3.host96{1,3}.cpu.1.*", ExpectedResults: 16},
{Pattern: "*.dc3.{host,server}96{1,3}.cpu.1.*", ExpectedResults: 16},
{Pattern: "*.dc3.{host,server}*{1,3}.cpu.1.*", ExpectedResults: 160},
```

Results:
```
name \ time/op     run1.noexpand.txt  run1.expandexact.txt  run1.fullexpand.txt
Find-8                    115µs ± 5%             68µs ± 1%           138µs ±32%
Concurrent4Find-8        65.2µs ± 6%           47.1µs ±10%          77.4µs ±12%
Concurrent8Find-8        64.2µs ± 4%           51.7µs ±10%          69.1µs ±22%

name \ alloc/op    run1.noexpand.txt  run1.expandexact.txt  run1.fullexpand.txt
Find-8                   42.5kB ± 0%           36.4kB ± 0%          69.3kB ± 0%
Concurrent4Find-8        42.5kB ± 0%           36.4kB ± 0%          69.3kB ± 0%
Concurrent8Find-8        42.5kB ± 0%           36.4kB ± 0%          69.3kB ± 0%

name \ allocs/op   run1.noexpand.txt  run1.expandexact.txt  run1.fullexpand.txt
Find-8                      215 ± 0%              160 ± 0%             210 ± 0%
Concurrent4Find-8           215 ± 0%              160 ± 0%             210 ± 0%
Concurrent8Find-8           215 ± 0%              160 ± 0%             210 ± 0%
```

So, option (2) seems to be the way to go for these queries.