@shanson7 2017-07-17
The downside of option (2) in the above case is that we have 2 different methods of expanding `{}` dependent on if there is a regex char in the string. I can't think of a clean and efficient method of deduplicating the parsing code (since one creates multiple strings and the other makes a regex). 

With this in mind, and @Dieterbe suggestion to reduce the scope of the changes, I'll go with option (3) to expand the queries and then operate on the multiple expansions.