@shanson7 2017-07-17
Ok, this last round of changes implemented scheme 3. Final benchmarks are:

With debug disabled (what prod would be in ideal case):
```
$ benchstat  master.txt fullExpand.txt
name               old time/op    new time/op    delta
Find-8                168µs ± 3%      76µs ±12%  -55.05%  (p=0.000 n=20+20)
Concurrent4Find-8    77.8µs ± 8%    39.8µs ± 9%  -48.82%  (p=0.000 n=19+19)
Concurrent8Find-8    68.4µs ±14%    37.5µs ±23%  -45.14%  (p=0.000 n=20+19)

name               old alloc/op   new alloc/op   delta
Find-8               51.2kB ± 0%    23.0kB ± 0%  -55.15%  (p=0.000 n=17+18)
Concurrent4Find-8    51.2kB ± 0%    23.0kB ± 0%  -55.15%  (p=0.000 n=17+20)
Concurrent8Find-8    51.2kB ± 0%    23.0kB ± 0%  -55.15%  (p=0.000 n=19+19)

name               old allocs/op  new allocs/op  delta
Find-8                  626 ± 0%       156 ± 0%  -75.08%  (p=0.000 n=20+20)
Concurrent4Find-8       626 ± 0%       156 ± 0%  -75.08%  (p=0.000 n=20+20)
Concurrent8Find-8       626 ± 0%       156 ± 0%  -75.08%  (p=0.000 n=20+20)
```

With debug enabled (what prod is in current case):
```
$ benchstat  master.dbg.txt fullExpand.dbg.txt
name               old time/op    new time/op    delta
Find-8                163µs ± 9%      93µs ± 8%  -43.08%  (p=0.000 n=19+20)
Concurrent4Find-8    70.4µs ± 9%    46.6µs ± 7%  -33.75%  (p=0.000 n=19+20)
Concurrent8Find-8    72.0µs ± 4%    40.0µs ± 5%  -44.49%  (p=0.000 n=19+19)

name               old alloc/op   new alloc/op   delta
Find-8               58.1kB ± 0%    28.7kB ± 0%  -50.68%  (p=0.000 n=20+19)
Concurrent4Find-8    58.1kB ± 0%    28.7kB ± 0%  -50.68%  (p=0.000 n=16+19)
Concurrent8Find-8    58.1kB ± 0%    28.7kB ± 0%  -50.68%  (p=0.000 n=20+20)

name               old allocs/op  new allocs/op  delta
Find-8                1.19k ± 0%     0.65k ± 0%  -45.85%  (p=0.000 n=20+20)
Concurrent4Find-8     1.19k ± 0%     0.65k ± 0%  -45.85%  (p=0.000 n=20+20)
Concurrent8Find-8     1.19k ± 0%     0.65k ± 0%  -45.85%  (p=0.000 n=20+20)
```