@Dieterbe 2017-08-11
hey @shanson7 this looks good now. i rebased them (so history will look cleaner) and merged into master. thanks!